+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { getSession } from "../lib/auth/getSession";
+import { SignOutButton } from "./SingOutButton";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -34,9 +35,7 @@ export default async function RootLayout({
           <h1 className="text-xl font-bold">DM Notes</h1>
           <div className="flex items-center gap-4">
             {session && <span>Logged in as: {session.login}</span>}
-            <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded">
-              Logout
-            </button>
+            <SignOutButton />
           </div>
         </header>
