+++ src/lib/auth/signIn.ts
@@ -1,25 +1,9 @@
 "use server";
 
-import { createUser, getUserByLogin } from "../repos/user";
+import { getUserByLogin } from "../repos/user";
 import { TOTP } from "totp-generator";
 import { redirect } from "next/navigation";
 import { getSessionWithMeta } from "./getSession";
-
-export async function verifyLogin({
-  login,
-}: {
-  login: string;
-}): Promise<{ loginExists: true } | { loginExists: false; secret: string }> {
-  const user = await getUserByLogin(login);
-
-  if (user) {
-    return { loginExists: true };
-  }
-
-  const newUser = await createUser(login);
-
-  return { loginExists: false, secret: newUser.OTPSecret };
-}
 
 export async function signIn({ login, otp }: { login: string; otp: string }) {
   const user = await getUserByLogin(login);
@@ -31,7 +15,7 @@ export async function signIn({ login, otp }: { login: string; otp: string }) {
   const isVerified = TOTP.generate(user.OTPSecret).otp === otp;
 
   if (!isVerified) {
-    throw new Error("Invalid OTP");
+    return { error: "Invalid OTP" };
   }
 
   const session = await getSessionWithMeta();
@@ -42,13 +26,6 @@ export async function signIn({ login, otp }: { login: string; otp: string }) {
   };
 
   await session.save();
-
-  redirect("/");
-}
-
-export async function logout() {
-  const session = await getSessionWithMeta();
-  session.destroy();
 
   redirect("/");
 }
